initial version of eventfd support (commented out for now)
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@825 7451a71c-b6c8-49cc-8254-141fc5361860

+++ libzmq/pipe.cpp
@@ -73,7 +73,8 @@ bool zmq::pipe_t::check_write ()
     //  The else branch will come into effect after 500,000 years of
     //  passing 1,000,000 messages a second but still, it's implemented just
     //  in case ...
-    int size = last_head <= tail ? tail - last_head :
+    //  TODO: Can size be possibly negative? If so, what then?
+    uint64_t size = last_head <= tail ? tail - last_head :
         std::numeric_limits <uint64_t>::max () - last_head + tail + 1;
     assert (size <= hwm);
